pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rexia-ai"
version = "0.1.0"
description = "Your package description"
authors = [
    { name = "Robyn Le Sueur", email = "robyn.lesueur@googlemail.com" }
]
dependencies = [
    "python>=3.8",
    "selenium>=3.141.0",
    "webdriver-manager>=3.4.2",
    "python-dotenv>=0.19.1",
    "python-docx>=0.8.11",
    "llama-index @ git+https://github.com/your-git-repo/llama-index.git@main"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5"
]
